Agregando validaciones al ABM de usuarios y al Generar Reserva

+++ src/FrbaHotel/Views/ABM de Usuario/ABMUsuario.cs
@@ -9,6 +9,7 @@ using System.Windows.Forms;
 
 using System.Data.SqlClient;
 
+using FrbaHotel.Homes;
 using FrbaHotel.Models;
 using FrbaHotel.Database_Helper;
 
@@ -52,11 +53,22 @@ namespace FrbaHotel.ABM_de_Usuario
 
         private void button2_Click(object sender, EventArgs e)
         {
+            List<RolUsuario> roles = (dataGridView1.SelectedRows[0].DataBoundItem as Usuario).getRoles();
+            if( !roles.Any(r => r.hotel.id == SesionActual.rol_usuario.hotel.id) ) {
+                MessageBox.Show("No se puede eliminar un usuario si no trabaja en el mismo hotel que el Administrador");
+                return;
+            }
             this.deleteRecord(dataGridView1);
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
+            List<RolUsuario> roles = (dataGridView1.SelectedRows[0].DataBoundItem as Usuario).getRoles();
+            if (!roles.Any(r => r.hotel.id == SesionActual.rol_usuario.hotel.id))
+            {
+                MessageBox.Show("No se puede modificar un usuario si no trabaja en el mismo hotel que el Administrador");
+                return;
+            }
             this.editRecord<Usuario, FrbaHotel.ABM_de_Usuario.AltaModificaiconUsuario>(dataGridView1);
         }
